requirements.txt
fastapi
uvicorn
sqlalchemy
requests
paho-mqtt
psutil
PyQt6
python-dotenv
